A first attempt at another multi-level bounding box calculation.  Since it is pre-computed at a higher level of detail, it may be possible to reduce the number of on-the-fly calculations that are being done now.
There are a few noticable bugs.  The bounding boxes are being drawn in the wrong places, which means that collisions are not really working, either. There is a problem with the way that the controls are being read, so it is not possible to control the characters at the moment!

In the end, the only method that will really work is enqueing the triangles in an octree...  I'll see if I can find an algorithm on the internet, or in one of my books.