Tried to fix some potential problems if SDL_mixer fails to start. This used to cause program crashes and a bunch of other problems when music was on.
It seems that we *should not* use .ogg files for anything other than background music at this time. The only way to truly get them to work would be to write a utility that decompresses the .ogg file in memory and then plays it as a wav. It will not crash the program and I think my workaround will somewhat work, but generally it is a bad idea...