Console: Temporary fix for rendering issues
As subsequent rendering operations do not set the renderer state properly when
it has been altered by the console rendering operation, the console rendering
operation leaves the rendering state in a suitable default state. This is a
temporary fix, every rendering operation on its own shall ensure that the
renderer state is set propery.